Fix int overflow on large torrents
svn path=/branches/bitsharp-0.50/; revision=114047

+++ src/MonoTorrent/MonoTorrent.Client/ClientEngine.cs
@@ -280,7 +280,7 @@ namespace MonoTorrent.Client
         {
             get
             {
-                return Toolbox.Accumulate<TorrentManager>(torrents, delegate(TorrentManager m) { return m.Monitor.DownloadSpeed; });
+                return (int)Toolbox.Accumulate<TorrentManager>(torrents, delegate(TorrentManager m) { return m.Monitor.DownloadSpeed; });
             }
         }
 
@@ -288,7 +288,7 @@ namespace MonoTorrent.Client
         {
             get
             {
-                return Toolbox.Accumulate<TorrentManager>(torrents, delegate(TorrentManager m) { return m.Monitor.UploadSpeed; });
+                return (int)Toolbox.Accumulate<TorrentManager>(torrents, delegate(TorrentManager m) { return m.Monitor.UploadSpeed; });
             }
         }
